backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
apscheduler==3.10.4
python-multipart==0.0.20